refactor: normalize file naming to kebab-case
