+++ src/components/game/handle-round-end.ts
@@ -1,6 +1,6 @@
 import { calculateScore } from '../../app/state/actions';
 import { toggleButtonsStatement } from './toggle-buttons-statement';
-import { updateScore } from './updateScore';
+import { updateScore } from './update-score';
 
 export function handleRoundEnd(points: number) {
   calculateScore(points);
