+++ src/components/game/update-score.ts
