+++ src/pages/practice/practice.ts
@@ -1,7 +1,7 @@
 import { saveTopicQuestions } from '../../app/state/actions';
 import { getState } from '../../app/state/store';
 import { showNextQuestion } from '../../components/game/show-next-question';
-import { updateScore } from '../../components/game/updateScore';
+import { updateScore } from '../../components/game/update-score';
 import { createLoadingView } from '../../components/ui/loading/loading';
 import { createPracticeCard } from '../../components/ui/practice-card/practice-card';
 import { createSidePanel } from '../../components/ui/practice-card/side-panel/side-panel';
